add get and delete endpoints for all item types
